api_gateway/requirements.txt
aniso8601==9.0.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-GraphQL
Flask-JWT-Extended==4.6.0
graphene
graphql-core
graphql-relay
graphql-server-core
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
migrate==0.3.8
promise==2.3
PyJWT==2.9.0
requests==2.32.3
Rx==1.6.3
six==1.16.0
urllib3==2.2.3
Werkzeug==3.0.4
